core restructuring
